+++ dcb-hub-admin-ui/src/helpers/filters.ts
@@ -1,0 +1,12 @@
+import { getGridStringOperators } from "@mui/x-data-grid-pro";
+
+export const standardFilters = getGridStringOperators().filter(({ value }) =>
+	[
+		"equals",
+		"contains" /* add more over time as we build in support for them */,
+	].includes(value),
+);
+
+export const equalsOnly = getGridStringOperators().filter(({ value }) =>
+	["equals"].includes(value),
+);