+++ dcb-hub-admin-ui/src/models/HostLMS.ts
@@ -1,7 +1,9 @@
 export interface HostLMS {
-	clientConfig: any;
 	id: string;
 	code: string;
 	name: string;
 	lmsClientClass: string;
+	clientConfig: any;
+	itemSuppressionRulesetName: string;
+	suppressionRulesetName: string;
 }