+++ dcb-hub-admin-ui/src/pages/agencies/[agencyId].tsx
@@ -65,15 +65,15 @@ export default function AgencyDetails({ agencyId }: AgencyDetails) {
 					title={t("ui.error.cannot_retrieve_record")}
 					message={t("ui.info.connection_issue")}
 					description={t("ui.info.try_later")}
-					action={t("ui.info.go_back")}
+					action={t("ui.action.go_back")}
 					goBack="/agencies"
 				/>
 			) : (
 				<Error
-					title={t("ui.error.record_not_found")}
-					message={t("ui.info.record_unavailable")}
-					description={t("ui.action.check_url")}
-					action={t("ui.info.go_back")}
+					title={t("ui.error.cannot_find_record")}
+					message={t("ui.error.invalid_UUID")}
+					description={t("ui.info.check_address")}
+					action={t("ui.action.go_back")}
 					goBack="/agencies"
 				/>
 			)}
@@ -97,7 +97,7 @@ export default function AgencyDetails({ agencyId }: AgencyDetails) {
 				<Grid xs={2} sm={4} md={4}>
 					<Stack direction={"column"}>
 						<Typography variant="attributeTitle">
-							{t("details.agency_id")}
+							{t("details.agency_uuid")}
 						</Typography>
 						<RenderAttribute attribute={agency?.id} />
 					</Stack>