+++ dcb-hub-admin-ui/src/pages/agencies/index.tsx
@@ -63,10 +63,11 @@ const Agencies: NextPage = () => {
 						},
 						{
 							field: "id",
-							headerName: "Agency ID",
+							headerName: "Agency UUID",
 							minWidth: 100,
 							flex: 0.5,
 							filterOperators,
+							filterable: false,
 						},
 						{
 							field: "code",