+++ dcb-hub-admin-ui/src/pages/bibs/[bibId].tsx
@@ -91,15 +91,15 @@ export default function SourceBibDetails({ bibId }: BibDetails) {
 					title={t("ui.error.cannot_retrieve_record")}
 					message={t("ui.info.connection_issue")}
 					description={t("ui.info.try_later")}
-					action={t("ui.info.go_back")}
+					action={t("ui.action.go_back")}
 					goBack="/bibs"
 				/>
 			) : (
 				<Error
-					title={t("ui.error.record_not_found")}
-					message={t("ui.info.record_unavailable")}
-					description={t("ui.action.check_url")}
-					action={t("ui.info.go_back")}
+					title={t("ui.error.cannot_find_record")}
+					message={t("ui.error.invalid_UUID")}
+					description={t("ui.info.check_address")}
+					action={t("ui.action.go_back")}
 					goBack="/bibs"
 				/>
 			)}
@@ -120,7 +120,7 @@ export default function SourceBibDetails({ bibId }: BibDetails) {
 				<Grid xs={2} sm={4} md={4}>
 					<Stack direction={"column"}>
 						<Typography variant="attributeTitle">
-							{t("details.source_bib_id")}
+							{t("details.source_bib_uuid")}
 						</Typography>
 						<RenderAttribute attribute={bib?.id} />
 					</Stack>
@@ -144,7 +144,7 @@ export default function SourceBibDetails({ bibId }: BibDetails) {
 				<Grid xs={2} sm={4} md={4}>
 					<Stack direction={"column"}>
 						<Typography variant="attributeTitle">
-							{t("details.source_system_id")}
+							{t("details.source_system_uuid")}
 						</Typography>
 						<RenderAttribute attribute={bib?.sourceSystemId} />
 					</Stack>
@@ -160,7 +160,7 @@ export default function SourceBibDetails({ bibId }: BibDetails) {
 				<Grid xs={2} sm={4} md={4}>
 					<Stack direction={"column"}>
 						<Typography variant="attributeTitle">
-							{t("details.contributor_id")}
+							{t("details.contributor_uuid")}
 						</Typography>
 						<RenderAttribute attribute={bib?.contributesTo?.id} />
 					</Stack>