+++ dcb-hub-admin-ui/src/pages/patronRequests/audits/[auditId].tsx
@@ -67,15 +67,15 @@ export default function AuditDetails({ auditId }: AuditDetails) {
 					title={t("ui.error.cannot_retrieve_record")}
 					message={t("ui.info.connection_issue")}
 					description={t("ui.info.try_later")}
-					action={t("ui.info.go_back")}
+					action={t("ui.action.go_back")}
 					goBack={goBackLink}
 				/>
 			) : (
 				<Error
 					title={t("ui.error.record_not_found")}
-					message={t("ui.info.record_unavailable")}
-					description={t("ui.action.check_url")}
-					action={t("ui.info.go_back")}
+					message={t("ui.error.invalid_UUID")}
+					description={t("ui.info.check_address")}
+					action={t("ui.action.go_back")}
 					goBack={goBackLink}
 				/>
 			)}
@@ -90,7 +90,7 @@ export default function AuditDetails({ auditId }: AuditDetails) {
 				<Grid xs={2} sm={4} md={4}>
 					<Stack direction={"column"}>
 						<Typography variant="attributeTitle">
-							{t("details.audit_id")}
+							{t("details.audit_uuid")}
 						</Typography>
 						<RenderAttribute attribute={audit?.id} />
 					</Stack>
@@ -130,7 +130,7 @@ export default function AuditDetails({ auditId }: AuditDetails) {
 				<Grid xs={2} sm={4} md={4}>
 					<Stack direction={"column"}>
 						<Typography variant="attributeTitle">
-							{t("details.patron_request_id")}
+							{t("details.patron_request_uuid")}
 						</Typography>
 						<RenderAttribute attribute={audit?.patronRequest?.id} />
 					</Stack>